backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
langchain>=0.2.0
langgraph==0.6.2
cerebras-cloud-sdk
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.10.0
redis==5.0.1
celery==5.3.4
requests==2.31.0